MAINT: Move methods from fluid solver scheme to superclass
