+++ src/FTHM_Solver/full_petsc_solver.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-
+from functools import cached_property
 from dataclasses import dataclass
 from typing import Any, Callable, Optional
 
@@ -107,6 +107,50 @@ class SolverScheme(ABC):
 
     def _register_equation_variable_groups(self):
         pass
+
+    @cached_property
+    def variable_groups(self) -> list[list[int]]:
+        """Assign the following groups to the variables:
+
+        0: Fluid pressure on all subdomains.
+        1: Interface darcy fluxes on all interfaces.
+
+        """
+        return FTHM_Solver.get_variables_group_ids(
+            model=self.model,
+            md_variables_groups=self._variable_groups_keys,
+        )
+
+    @cached_property
+    def equation_groups(self) -> list[list[int]]:
+        """Define the groups of equation in the specific order, that we will use in
+        the block Jacobian to access the submatrices.
+
+        Returns:
+            List of lists of integers. Each list contains the indices of the equations
+                in the group.
+
+        """
+        return FTHM_Solver.get_equations_group_ids(
+            model=self.model,
+            equations_group_order=self._equation_group_keys,
+        )
+
+    def _group_id_from_name(self, name: str) -> int:
+        """Get the group id from the name of the group.
+
+        Args:
+            name: Name of the group.
+
+        Returns:
+            Group id.
+
+        """
+        # This is rough, does not allow for duplicates.
+        for i, group in enumerate(self._equation_group_keys):
+            if group[0][0] == name:
+                return [i]
+        raise ValueError(f"Group {name} not found.")
 
 
 class PreconditionerScheme(ABC):
